refactor(worker): split auth routes from worker
